refactored

+++ helper.py
@@ -7,12 +7,25 @@ class Helper:
         self.stemming_directory = "stemming/"
         self.lemmatization_directory = "lemmatization/"
 
-    def read_file(self, filename):
-        with open(self.directory_original_books + filename, "r") as f:
+    def get_directory(self, identifier):
+        directory = self.directory_original_books
+
+        if identifier == "f":
+            directory = self.directory_filtered_book
+        elif identifier == "s":
+            directory = self.stemming_directory
+        elif identifier == "l":
+            directory = self.lemmatization_directory
+
+        return directory
+
+    def read_file(self, filename, identifier=""):
+        directory = self.get_directory(identifier)
+        with open(directory + filename, "r") as f:
             return unicode(f.read(), errors="ignore")
 
-    def write_file(self, book, filename, dest=""):
-        directory = self.directory_original_books
+    def write_file(self, book, filename, identifier=""):
+        directory = self.get_directory(identifier)
 
         if not os.path.exists(self.directory_original_books):
             os.makedirs(self.directory_original_books)
@@ -23,15 +36,13 @@ class Helper:
         if not os.path.exists(self.lemmatization_directory):
             os.makedirs(self.lemmatization_directory)
 
-        if dest == "f":
-            directory = self.directory_filtered_book
-        elif dest == "s":
-            directory = self.stemming_directory
-        elif dest == "l":
-            directory = self.lemmatization_directory
+        if not os.path.exists(directory + filename):
+            with open(directory + filename, "w") as f:
+                print "Writing " + filename
+                f.write(book)
 
-        with open(directory + filename, "w") as f:
-            print "Writing " + filename
-            f.write(book)
-
-        print "Book successfully written to", directory + filename
+            print "Book successfully written to", directory + filename
+            return True
+        else:
+            print directory + filename, "already exists"
+            return False
